+++ formula-android/src/main/java/com/instacart/formula/android/BaseFormulaFragment.kt
@@ -1,6 +1,4 @@
-package com.instacart.formula.fragment
-
-import com.instacart.formula.android.FragmentKey
+package com.instacart.formula.android
 
 interface BaseFormulaFragment<State> {