+++ formula-android/src/main/java/com/instacart/formula/fragment/FormulaFragment.kt
@@ -6,6 +6,7 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import com.instacart.formula.Cancelable
+import com.instacart.formula.android.BaseFormulaFragment
 import com.instacart.formula.android.FeatureView
 import com.instacart.formula.android.FragmentEnvironment
 import com.instacart.formula.android.FragmentKey